rest api changes for project and task and delete user

+++ src/main/java/com/cts/fse/projectmanager/services/UIService.java
@@ -20,11 +20,14 @@ import com.cts.fse.projectmanager.dto.Task;
 import com.cts.fse.projectmanager.dto.TaskCreateDTO;
 import com.cts.fse.projectmanager.dto.TaskUI;
 import com.cts.fse.projectmanager.dto.UIUsers;
+import com.cts.fse.projectmanager.dto.UserProject;
 import com.cts.fse.projectmanager.dto.Users;
 import com.cts.fse.projectmanager.repository.ParentTaskRepository;
 import com.cts.fse.projectmanager.repository.ProjectRepository;
 import com.cts.fse.projectmanager.repository.TaskRepository;
+import com.cts.fse.projectmanager.repository.UserProjectRepository;
 import com.cts.fse.projectmanager.repository.UsersRepository;
+import com.cts.fse.projectmanager.utils.InvalidRequestException;
 
 @Service
 public class UIService implements IUIService {
@@ -40,19 +43,22 @@ public class UIService implements IUIService {
 
 	@Autowired
 	private ParentTaskRepository parentTaskRepository;
+	
+	@Autowired
+	private UserProjectRepository userProjectRepository;
 
 	@Override
-	public boolean createUser(Users user) {
+	public boolean createUser(Users user) throws InvalidRequestException{
 		try {
 			usersRepository.save(user);
 			return true;
 		} catch (Exception e) {
-			return false;
+			throw new InvalidRequestException("User is not saved in database");
 		}
 	}
 
 	@Override
-	public boolean updateUser(Users user, long id) {
+	public boolean updateUser(Users user, long id) throws InvalidRequestException{
 		try {
 			Optional<Users> userOpt = usersRepository.findById(id);
 			if (userOpt.isPresent()) {
@@ -64,23 +70,24 @@ public class UIService implements IUIService {
 				return true;
 			}
 		} catch (Exception e) {
-			return false;
+			throw new InvalidRequestException("User is not updated in database");
 		}
 		return false;
 	}
 
 	@Override
-	public Users findUserById(long id) {
+	public Users findUserById(long id) throws InvalidRequestException{
 		Optional<Users> userOpt = usersRepository.findById(id);
 		if (userOpt.isPresent()) {
 			return userOpt.get();
 		} else {
-			return null;
+			throw new InvalidRequestException("User is not found in database");
 		}
 	}
 
 	@Override
-	public List<UIUsers> findAllUsers() {
+	public List<UIUsers> findAllUsers() throws InvalidRequestException{
+		try {
 		List<Users> usersList = usersRepository.findAll();
 		List<UIUsers> listUsers = new ArrayList<>();
 		usersList.stream().forEach(user -> {
@@ -92,15 +99,25 @@ public class UIService implements IUIService {
 			listUsers.add(users);
 		});
 		return listUsers;
+		}catch(Exception e) {
+			throw new InvalidRequestException("Exception while fetching all users :"+e.getLocalizedMessage());
+		}
 	}
 
 	@Override
-	public void deleteUserByid(long id) {
-		usersRepository.deleteById(id);
+	public void deleteUserByid(long id) throws InvalidRequestException {
+		List<UserProject> userProjectList = userProjectRepository.getUserProject(id);
+		if (!userProjectList.isEmpty())
+			throw new InvalidRequestException("User is mapped with project,can not delete user");
+		try {
+			usersRepository.deleteById(id);
+		} catch (Exception e) {
+			throw new InvalidRequestException("User is not found in database");
+		}
 	}
 
 	@Override
-	public boolean createProject(ProjectCreateDTO projectCreateDTO) {
+	public boolean createProject(ProjectCreateDTO projectCreateDTO) throws InvalidRequestException{
 		try {
 			Project project = new Project();
 			project.setPriority(projectCreateDTO.getPriority());
@@ -117,12 +134,12 @@ public class UIService implements IUIService {
 			usersRepository.save(user);
 			return true;
 		} catch (Exception e) {
-			return false;
+			throw new InvalidRequestException("Project is not saved in database");
 		}
 	}
 
 	@Override
-	public boolean createTask(TaskCreateDTO taskCreateDTO) {
+	public boolean createTask(TaskCreateDTO taskCreateDTO) throws InvalidRequestException{
 		try {
 			if ("true".equals(taskCreateDTO.getIsParentTask())) {
 				ParentTask parent = new ParentTask();
@@ -159,12 +176,13 @@ public class UIService implements IUIService {
 				return true;
 			}
 		} catch (Exception e) {
-			return false;
+			throw new InvalidRequestException("Task is not saved in database");
 		}
 	}
 
 	@Override
-	public List<ProjectDataTableDTO> findAllProjects() {
+	public List<ProjectDataTableDTO> findAllProjects() throws InvalidRequestException{
+		try {
 		List<ProjectDataTableDTO> tableList = new ArrayList<>();
 		SimpleDateFormat expectedformatter = new SimpleDateFormat("yyyy-MM-dd");
 		List<Project> projectList = projectRepository.findAll();
@@ -191,14 +209,17 @@ public class UIService implements IUIService {
 			tableList.add(projectDataTableDTO);
 		});
 		return tableList;
+		}catch(Exception e) {
+			throw new InvalidRequestException("Exception while fetching all projects:"+e.getLocalizedMessage());
+		}
 	}
 
 	@Override
-	public boolean updateProject(ProjectCreateDTO projectCreateDTO, long id) {
+	public boolean updateProject(ProjectCreateDTO projectCreateDTO, long id) throws InvalidRequestException{
 		try {
 			Optional<Project> projectOpt = projectRepository.findById(id);
 			if (projectOpt.isPresent()) {
-				Project project = new Project();
+				Project project =projectOpt.get();
 				project.setPriority(projectCreateDTO.getPriority());
 				project.setEndDate(projectCreateDTO.getEndDate());
 				project.setStartDate(projectCreateDTO.getStartDate());
@@ -214,13 +235,14 @@ public class UIService implements IUIService {
 				return true;
 			}
 		} catch (Exception e) {
-			return false;
+			throw new InvalidRequestException("Project is not saved in database");
 		}
 		return false;
 	}
 
 	@Override
-	public List<ParentTaskUI> findAllParentTask() {
+	public List<ParentTaskUI> findAllParentTask() throws InvalidRequestException{
+		try {
 		List<ParentTask> parentTaskList = parentTaskRepository.findAll();
 		List<ParentTaskUI> parentTaskUIList = new ArrayList<>();
 		parentTaskList.forEach(parenTask -> {
@@ -230,10 +252,14 @@ public class UIService implements IUIService {
 			parentTaskUIList.add(parentTaskUI);
 		});
 		return parentTaskUIList;
+		}catch(Exception e) {
+			throw new InvalidRequestException("Exception while fetching all parent task:"+e.getLocalizedMessage());
+		}
 	}
 
 	@Override
-	public List<ProjectTaskDTO> findAllProjectTaskMap() {
+	public List<ProjectTaskDTO> findAllProjectTaskMap() throws InvalidRequestException{
+		try {
 		List<ProjectTaskDTO> allProjectTask = new ArrayList<>();
 		List<Task> allTask = taskRepository.findAll();
 		allTask.forEach(task -> {
@@ -243,10 +269,14 @@ public class UIService implements IUIService {
 			allProjectTask.add(projectTaskDTO);
 		});
 		return allProjectTask;
+		}catch(Exception e) {
+			throw new InvalidRequestException("Exception while fetching all ProjectTaskMap:"+e.getLocalizedMessage());
+		}
 	}
 
 	@Override
-	public List<TaskUI> findAllTaskByProjectId(long id) {
+	public List<TaskUI> findAllTaskByProjectId(long id) throws InvalidRequestException{
+		try {
 		List<TaskUI> taskUIList=new ArrayList<TaskUI>();
 		SimpleDateFormat expectedformatter = new SimpleDateFormat("yyyy-MM-dd");
 		List<Task> allTask = taskRepository.findAllTaskByProjectId(id);
@@ -273,19 +303,24 @@ public class UIService implements IUIService {
 			taskUIList.add(taskUI);
 		});
 		return taskUIList;
+		}catch(Exception e) {
+			throw new InvalidRequestException("Exception while fetching all task:"+e.getLocalizedMessage());
+		}
 	}
 
 	@Override
-	public boolean updateTask(TaskCreateDTO taskCreateDTO, long id) {
+	public boolean updateTask(TaskCreateDTO taskCreateDTO, long id) throws InvalidRequestException{
 		try {
 			if ("true".equals(taskCreateDTO.getIsParentTask())) {
+				Optional<ParentTask> parentOpt=parentTaskRepository.findById(taskCreateDTO.getParentId());
+				ParentTask parent=parentOpt.get();
+				parent.setParentTask(taskCreateDTO.getTask());
+				parent = parentTaskRepository.save(parent);
 				Optional<Task> taskOpt = taskRepository.findById(id);
 				Task task=taskOpt.get();
 				Optional<Users> userOpt = usersRepository.findById(taskCreateDTO.getUserId());
 				Users user = userOpt.get();
 				task.setUsers(user);
-				Optional<ParentTask> parentOpt=parentTaskRepository.findById(taskCreateDTO.getParentId());
-				ParentTask parent=parentOpt.get();
 				Optional<Project> projectOpt = projectRepository.findById(taskCreateDTO.getProjectId());
 				Project project = projectOpt.get();
 				task.setProject(project);
@@ -312,7 +347,7 @@ public class UIService implements IUIService {
 				return true;
 			}
 		} catch (Exception e) {
-			return false;
+			throw new InvalidRequestException("Task is not saved in database");
 		}
 	}
